feat: 시간표 관련 로직 분리 & 리팩토링
컴포넌트를 분리 & 로직 또한 컴포넌트에서 분리
